Renamed packages in accordance with proper naming convention
Some of the packages had capital letters in their names which I removed.
