+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/RightSpecimenAuto.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opModes.Auto;
+package org.firstinspires.ftc.teamcode.opmodes.auto;
 
 import com.acmerobotics.roadrunner.ParallelAction;
 import com.acmerobotics.roadrunner.Pose2d;
